requirements.txt
tls-client
websocket-client
solders
base58
